// src/app/interfaces/ipool.ts
export interface Ipool {
  type: string;
  address: string;
  stakedTokenImg: string;
  subsidiaryTokenImg1: string;
  subsidiaryTokenImg2: string;
  rewardTokenImg: string;
  multiplier: number;
  factory: string;
  farm: string;
  subsidiaryTokenSymbol1: string;
  subsidiaryTokenSymbol2: string;
}